Added a few functions, moved CollisionInfo outside of CourseColMgr, started ObjectColMgr
