+++ source/game/field/CollisionDirector.hh
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "game/field/CourseColMgr.hh"
+#include "game/field/CollisionInfo.hh"
 
 #include <egg/math/Vector.hh>
 
@@ -20,17 +21,17 @@ public:
             u32 /*unused*/);
 
     bool checkSphereFull(f32 radius, const EGG::Vector3f &v0, const EGG::Vector3f &v1,
-            KCLTypeMask flags, CourseColMgr::CollisionInfo *pInfo, KCLTypeMask *pFlagsOut,
+            KCLTypeMask flags, CollisionInfo *pInfo, KCLTypeMask *pFlagsOut,
             u32 /*start*/);
     bool checkSphereFullPush(f32 radius, const EGG::Vector3f &v0, const EGG::Vector3f &v1,
-            KCLTypeMask flags, CourseColMgr::CollisionInfo *pInfo, KCLTypeMask *pFlagsOut,
+            KCLTypeMask flags, CollisionInfo *pInfo, KCLTypeMask *pFlagsOut,
             u32 /*start*/);
 
     bool checkSphereCachedPartialPush(const EGG::Vector3f &pos, const EGG::Vector3f &prevPos,
-            KCLTypeMask typeMask, CourseColMgr::CollisionInfo *colInfo, KCLTypeMask *typeMaskOut,
+            KCLTypeMask typeMask, CollisionInfo *colInfo, KCLTypeMask *typeMaskOut,
             f32 radius, u32 start);
     bool checkSphereCachedFullPush(const EGG::Vector3f &pos, const EGG::Vector3f &prevPos,
-            KCLTypeMask typeMask, CourseColMgr::CollisionInfo *colInfo, KCLTypeMask *typeMaskOut,
+            KCLTypeMask typeMask, CollisionInfo *colInfo, KCLTypeMask *typeMaskOut,
             f32 radius, u32 start);
 
     void resetCollisionEntries(KCLTypeMask *ptr);
