+++ source/game/field/CollisionInfo.cc
@@ -1,0 +1,41 @@
+#include "CollisionInfo.hh"
+
+#include "game/field/KCollisionTypes.hh"
+
+namespace Field
+{
+
+void CollisionInfo::updateFloor(f32 dist, const EGG::Vector3f &fnrm) {
+    if (dist > floorDist) {
+        floorDist = dist;
+        floorNrm = fnrm;
+    }
+}
+
+void CollisionInfo::updateWall(f32 dist, const EGG::Vector3f &fnrm) {
+    if (dist > wallDist) {
+        wallDist = dist;
+        wallNrm = fnrm;
+    }
+}
+
+void CollisionInfo::update(f32 now_dist, const EGG::Vector3f &offset,
+        const EGG::Vector3f &fnrm, u32 kclAttributeTypeBit) {
+    bbox.min = bbox.min.minimize(offset);
+    bbox.max = bbox.max.maximize(offset);
+
+    if (kclAttributeTypeBit & KCL_TYPE_FLOOR) {
+        updateFloor(now_dist, fnrm);
+    } else if (kclAttributeTypeBit & KCL_TYPE_WALL) {
+        if (wallDist > -std::numeric_limits<f32>::min()) {
+            f32 dot = 1.0f - wallNrm.ps_dot(fnrm);
+            if (dot > perpendicularity) {
+                perpendicularity = std::min(dot, 1.0f);
+            }
+        }
+
+        updateWall(now_dist, fnrm);
+    }
+}
+
+} // namespace Field
