+++ source/game/field/CourseColMgr.hh
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "game/field/CollisionInfo.hh"
 #include "game/field/KColData.hh"
 
 #include <egg/math/BoundBox.hh>
@@ -14,23 +15,6 @@ typedef bool (
 
 class CourseColMgr {
 public:
-    struct CollisionInfo {
-        EGG::BoundBox3f bbox;
-        EGG::Vector3f tangentOff;
-        EGG::Vector3f floorNrm;
-        EGG::Vector3f wallNrm;
-        EGG::Vector3f _3c;
-        f32 floorDist;
-        f32 wallDist;
-        f32 _50;
-        f32 perpendicularity;
-
-        void updateFloor(f32 dist, const EGG::Vector3f &fnrm);
-        void updateWall(f32 dist, const EGG::Vector3f &fnrm);
-        void update(f32 now_dist, const EGG::Vector3f &offset, const EGG::Vector3f &fnrm,
-                u32 kclAttributeTypeBit);
-    };
-
     struct NoBounceWallColInfo {
         EGG::BoundBox3f bbox;
         EGG::Vector3f tangentOff;
