+++ source/game/kart/KartCollide.hh
@@ -3,7 +3,7 @@
 #include "game/kart/CollisionGroup.hh"
 #include "game/kart/KartObjectProxy.hh"
 
-#include <game/field/CourseColMgr.hh>
+#include <game/field/CollisionInfo.hh>
 
 namespace Kart {
 
@@ -27,11 +27,11 @@ public:
             const EGG::Vector3f &center, f32 radius);
 
     void processWheel(CollisionData &collisionData, Hitbox &hitbox,
-            Field::CourseColMgr::CollisionInfo *colInfo, Field::KCLTypeMask *maskOut);
+            Field::CollisionInfo *colInfo, Field::KCLTypeMask *maskOut);
     void processBody(CollisionData &collisionData, Hitbox &hitbox,
-            Field::CourseColMgr::CollisionInfo *colInfo, Field::KCLTypeMask *maskOut);
+            Field::CollisionInfo *colInfo, Field::KCLTypeMask *maskOut);
     void processFloor(CollisionData &collisionData, Hitbox &hitbox,
-            Field::CourseColMgr::CollisionInfo *colInfo, Field::KCLTypeMask *maskOut, bool wheel);
+            Field::CollisionInfo *colInfo, Field::KCLTypeMask *maskOut, bool wheel);
 
     void applySomeFloorMoment(f32 down, f32 rate, CollisionGroup *hitboxGroup,
             const EGG::Vector3f &forward, const EGG::Vector3f &nextDir, const EGG::Vector3f &speed,
@@ -39,7 +39,7 @@ public:
 
     bool FUN_805B6A9C(CollisionData &collisionData, const Hitbox &hitbox, EGG::BoundBox3f &minMax,
             EGG::Vector3f &relPos, s32 &count, Field::KCLTypeMask &maskOut,
-            const Field::CourseColMgr::CollisionInfo &colInfo);
+            const Field::CollisionInfo &colInfo);
     void applyBodyCollision(CollisionData &collisionData, const EGG::Vector3f &movement,
             const EGG::Vector3f &posRel, s32 count);
